services/users/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2
MarkupSafe==1.1.1
Werkzeug==0.15.2
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.6.1
Flask-Testing==0.7.1
parameterized
gunicorn
coverage
flake8
flask-cors==3.0.7
flask-migrate
flask-bcrypt
pyjwt
